requirements.txt
Django==1.5.3
django-rest-framework-docs==0.1.7
django-rest-swagger==0.1.9
djangorestframework==2.3.8
jsonpickle==0.6.1
wsgiref==0.1.2
